requirements.txt
Flask
python-dotenv
google-generativeai
validators
markupsafe
